Simplify prompt token arithmetic
Cast the unsigned token count before multiplication and rely on the safe integer promotion of num_beams.

Files changed:
- src/telemetry/generation_telemetry.cpp: remove the redundant num_beams cast.

Copilot-Session: 6b269873-a7bf-490e-a412-70a920868135

+++ src/telemetry/generation_telemetry.cpp
@@ -69,8 +69,7 @@ void GenerationTelemetry::CompleteAppend(size_t input_token_count, int num_beams
   append_tracking_ = false;
 
   if (prompt_tokens_ == 0 && !first_token_logged_) start_time_ = append_start_time_;
-  prompt_tokens_ +=
-      static_cast<int64_t>(input_token_count) * static_cast<int64_t>(num_beams);
+  prompt_tokens_ += static_cast<int64_t>(input_token_count) * num_beams;
   input_modality_ = input_modality;
 }
